Fix yaml.R not being able to read rules files with three dashes
Yaml.R uses a regex to split a read yaml file in order to seperate the
options which can occur in a file from the rest of the file, which
should contain the rules. This regex was not strict enough, causing it
to also match on three dashes used in the middle of a rules expression.
This commit makes the regex stricter, fixing this issue.